GroupsOnGrids

Documentation for GroupsOnGrids.

GroupsOnGrids.C1

GroupsOnGrids.C2

GroupsOnGrids.C4

GroupsOnGrids.D1

GroupsOnGrids.D2

GroupsOnGrids.D4

GroupsOnGrids.AxesInfo

GroupsOnGrids.PointGroup

GroupsOnGrids.SpaceGroup

GroupsOnGrids.act_on_grid

GroupsOnGrids.adjust_axes

GroupsOnGrids.lift

GroupsOnGrids.swap_axes

GroupsOnGrids.C1 — Constant

The cyclic group of order 1.

This is the trivial group, which only contains the identity.

Elements

1: e, the identity

GroupsOnGrids.C2 — Constant

The cyclic group of order 2.

This is the group of rotations of a line segment, which contains the identity and a rotation by 180 degrees.

Elements

1: e, the identity

2: r2, a rotation by 180 degrees

GroupsOnGrids.C4 — Constant

The cyclic group of order 4.

This group is generated by a rotation by 90 degrees.

Elements

1: e, the identity

2: r4, a rotation by 90 degrees

3: r4^2, a rotation by 180 degrees

4: r4^3, a rotation by 270 degrees

GroupsOnGrids.D1 — Constant

The dihedral group of order 2.

This is the group of symmetries of a line segment, which contains the identity and a reflection.

Elements

1: e, the identity

2: s, a reflection

GroupsOnGrids.D2 — Constant

The dihedral group of order 4.

This group is generated by a reflection and a rotation by 180 degrees.

Elements

1: e, the identity

2: s, a reflection

3: r2 s, a reflection followed by a rotation by 180 degrees

4: r2, a rotation by 180 degrees

GroupsOnGrids.D4 — Constant

The dihedral group of order 8.

This group is generated by a reflection and a rotation by 90 degrees.

Elements

1: e, the identity

2: r4, a rotation by 90 degrees

3: r4^2, a rotation by 180 degrees

4: r4^3, a rotation by 270 degrees

5: s, a reflection

6: r4 s, a reflection followed by a rotation by 90 degrees

7: r4^2 s, a reflection followed by a rotation by 180 degrees

8: r4^3 s, a reflection followed by a rotation by 270 degrees

GroupsOnGrids.AxesInfo — Type

Represents the meaning of the axes of a signal.

Fields

width: The axis of the signal representing the width.

height: The axis of the signal representing the height.

depth: The axis of the signal representing the depth.

group: The axis of the signal representing the group element.

GroupsOnGrids.PointGroup — Type

A point group is a group that leaves at least one element fixed. Here, they are the part of the space groups that do not involve translations. This leaves rotations and reflections.

Fields

name::String: The name.

order::Int: The number of elements.

composition::Vector{Vector{Int}}: The composition table.

inverses::Vector{Int}: The inverses.

subgroups::Dict{String,Vector{Int}}: The subgroups.

GroupsOnGrids.SpaceGroup — Type

The space groups.

Each space group is defined by a point group and a way of using it to act on a grid.

GroupsOnGrids.act_on_grid — Method

Apply a space group to a grid.

Arguments

g: The group to apply.

x: The grid to act on.

axes: The meaning of the axes of the grid.

new_group_axis: The axis on which group actions should be concatenated.

Returns

The grid after the group action has been applied.

GroupsOnGrids.adjust_axes — Method

In case the group axis comes before the width or height axis, we need to adjust the axes.

GroupsOnGrids.lift — Method

Add a dimension to the signal representing the group element acted by.

The new group axis will bbe at the position new_group_axis, while the other axes maintain their relative order and are shifted accordingly.

GroupsOnGrids.swap_axes — Method

Swap two given axes, keeping the other axes in their relative order.