GroupsOnGrids

Documentation for GroupsOnGrids.

GroupsOnGrids.AxesInfo

GroupsOnGrids.GroupAction

GroupsOnGrids.PointGroup

GroupsOnGrids.act_on_grid

GroupsOnGrids.adjust_axes

GroupsOnGrids.lift

GroupsOnGrids.swap_axes

GroupsOnGrids.AxesInfo — Type

Represents the meaning of the axes of a signal.

Fields

width: The axis of the signal representing the width.

height: The axis of the signal representing the height.

depth: The axis of the signal representing the depth.

group: The axis of the signal representing the group element.

GroupsOnGrids.GroupAction — Type

Represents the action of all elements of a group.

Fields

group: The group that is acting.

new_group_axis: The position of the new group axis.

GroupsOnGrids.PointGroup — Type

A point group is a group that leaves at least one element fixed. Here, they are the part of the space groups that do not involve translations. This leaves rotations and reflections.

Fields

name::String: The name.

order::Int: The number of elements.

composition::Vector{Vector{Int}}: The composition table.

inverses::Vector{Int}: The inverses.

subgroups::Dict{String,Vector{Int}}: The subgroups.

GroupsOnGrids.act_on_grid — Method

Apply a group action to a grid.

Arguments

ga: The group action to apply.

x: The grid to act on.

axes: The meaning of the axes of the grid.

Returns

The grid after the group action has been applied.

GroupsOnGrids.adjust_axes — Method

In case the group axis comes before the width or height axis, we need to adjust the axes.

GroupsOnGrids.lift — Method

Add a dimension to the signal representing the group element acted by.

The new group axis will bbe at the position new_group_axis, while the other axes maintain their relative order and are shifted accordingly.

GroupsOnGrids.swap_axes — Method

Swap two given axes, keeping the other axes in their relative order.